series binding to expressions now works...
...but there are problems ahead wrt iterators. This is one of the places
where type erasure really hurts, because we don't just ask for iterators
and increment/decrement them - we also dereference them and hand them
off to others, and that's where a typed-iterator is problematic if we're
getting it from a non-typed frame/series. In frame::rows(), we need to
iterate the rows, evaluating the expression for each row. How do we get
a typed-iterator out of the old non-typed frame? We need a template type
list of course - except that the charm of our approach to rows() was
that it wouldn't need one? So perhaps we leave frame::iterator
untyped? That means that we wouldn't be able to dereference it in a nice
way without creating an untyped row class (because std::tuple is
stubbornly typed). The thing about C++ and typing is that when you erase
a type, eventually if you want to do interesting stuff with it you need
to add the types back, and without reification there's no way to do
that. Maybe that's obvious (d'uh, what would teh callsite look like
noob), but perhaps more frustrating, you can't "extend" the type erasure
with stuff like 'auto', because along with reification, C++ doesn't
support stuff like boxing.

So here's an aggregation tree

frame
  frame::iterator
  series
    series::iterator<T>=series_vector<T>::iterator  -- wut
    series_vector<T>
      series_vector<T>::iterator=std::vector<T>::iterator
      std::vector<T>

I think series::iterator should be untyped, I think? But then what does
it return with dereferencing? A void*? Is that in keeping with the
spirit of type-erasure? The dual of a series holding a series_vector<T>
through a iseries_vector would be a series::iterator holding a
series_vector<T>::iterator through an iseries_vector::iiterator... but
what's in that interface? How could this support C++'s interface idiom?

Maybe it doesn't? Could iseries_vector::iiterator support
increment/decrement, but not dereferencing? And while it clearly
wouldn't be a proper iterator, it would allow us to defer dereferencing
until it's held by a non-erased type like series_vector<T> or
binary_expr?