mf::series now has a type parameter - no longer type-erased.
series_vector still needs type-erasure with iseries_vector - this is how
mf::column <=> mf::series works.

So in general, the way I'm solving the typing problem is to make
everything typed (frame, series, series_vector). This ofc makes
frame::columns() way way more clumsy, if we refer to columns using
strings, because we need the return value (an mf::frame<Ts...>) to have
a type and we can't figure out what it would be in the signature of
frame::columns() in any easy way.
One alternative that I explored was using "symbols" (like in julia or
other languages), which in c++ would really be types, to refer to
columns. But then we'd need to declare the column symbols ahead of time
and it would be tricky to handle user data (like from a csv file) in any
nice way.
So instead, I'm creating a mf::columns class which is basically
mf::frame but without the column types. Then, mf::columns::frame<Ts...>
will create a dataframe with the requested types. Ideally it should try
to do a type conversion on the columns if <Ts...> is not the same as the
underlying series_vector<T> types. For example, we have

frame<double, int, string> foo;
foo.names() = { "temperature", "minutes", "label" };
auto bar = foo.columns( "minutes", "label" ).frame<float, string>();

...ideally we try to coerce the series_vector<int> which holds minutes
to a series_vector<float>. That way mf::columns::frame<Ts...>() isn't as
weird - it has a real purpose other than just making up for c++
deficiencies and restoring the frame types.