Changes to make frame::columns() work.
frame::columns() needs to return a frame_untyped because of the
limitations of referring to columns a runtime string - frame is typed
and a return value can't be typed based on runtime parameters, so we
need to go through the extra step of re-adding the types with
frame_untyped::to_frame<>(). We could add some extra functionality
perhaps by converting/coercing columns to different types if the erased
type isn't the same as the type specified. In that case, we'd probably
rename to_frame<> to types<> or something and actually implement this
method in frame<>::types<> as well.