Added the bitwise negation operator to mi
There are some issues to consider with logical negation and
missing-allowed columns. If we have

auto f1 = frame<mi<bool>, bool>{};
f1.push_back( false, true );
auto f2 = f1.new_series<mi<double>>( "third col", !_0 && _1 );

... what value does _2 get for these two rows?
_0 is false so !false is true so true && true is true
BUT
optional is engaged (not missing) so !_0 is false so false && true is false

The problem here is that std::optional already responds to operator!().
Or, it coerces to a bool that corresponds with "engaged". I'm not sure
which. But either way, it sort of interferes with the ability to test a
column for truth/falsehood. But also, in the case of

f1.push_back( missing, true );

...we'd really like a way of testing for missing in expressions. In
theory, we could completely steal operator!() away from mi/optional and
test for missing like

auto f2 = f1.new_series<mi<double>>( "third col", (_0 == missing) && _1 );