This should get the Windows msvc build up and running.
Getting the msvc build working has been useful - it exposed the
double->size_t coercion problem in my operator<<() functions, and while
I'm not 100% sure about how useful exposing the empty structs/classes
issue with expr_column was, the spec is the spec and it's good to
address the issue.

But there are some warnings that I'm getting out of msvc that are just
wrong.

There's the name-scoping/masking issue - I can understand that a dev
might accidentally declare a local with the same name as a global and
might be confused about strange behaviour or a compile error, but that's
sort of the whole point about having name scopes. Warning about use of a
legitimate feature of the language is probably wrong. Maybe someone
could point me to a really tricky case where this would be really
difficult to diagnose, but in the meantime, I'm disabling C4459.

But msvc C4702 warning is just completely broken with if constexpr ().
Originally I had

    bool any_missing_impl() const
    {
        const U& elem = at< Ind >();
        if constexpr ( is_missing<U>::value ) {
            if ( !elem.has_value() ) {
                return true;
            }
        }
        if constexpr ( sizeof...(Us) > 0 ) {
            return any_missing_impl<Ind + 1, Us...>();
        }
        return false;
    }

Clearly, the final "return false;" is my base-case and will get hit in
the case that the last column either isn't of the form mi<T> or is
mi<T> but isn't "missing".
But msvc complains that the "return false;" is unreachable code. Then,
when you change it to the technically correct but slightly less
attractive

    bool any_missing_impl() const
    {
        const U& elem = at< Ind >();
        if constexpr ( is_missing<U>::value ) {
            if ( !elem.has_value() ) {
                return true;
            }
        }
        else {
            if constexpr ( sizeof...(Us) > 0 ) {
                return any_missing_impl<Ind + 1, Us...>();
            }
            return false;
        }
    }

...it now complains that not all code paths return a value. So
perhaps the problem was the second if constexpr? I changed it to

    bool any_missing_impl() const
    {
        const U& elem = at< Ind >();
        if constexpr ( is_missing<U>::value ) {
            if ( !elem.has_value() ) {
                return true;
            }
        }
        if constexpr ( sizeof...(Us) > 0 ) {
            return any_missing_impl<Ind + 1, Us...>();
        }
        else { // VC++ C4702 is broken with if constexpr ()
            return false;
        }
    }

... and this works. But I don't want to change it to this - both of
those other versions are logically correct and the first is easier for
me to understand. So I'm disabling warning C4702 altogether.