grouped_frame<>::count()
I have the very outlines of grouped_frame<> worked out, kind of.
g_f::count() works because the extra column is a single int so it meant
that it worked well with frame::push_back(Row, Ts...) but it's harder to
understand how to make this work with the more free-form functions like
max(columnindex<Ind>...) or mean(columnindex<Ind>...).