frame::remove_column(), and some other changes
In addition to frame::remove_column(), I've renamed set_series and
add_series to set_column and add_column. From now on, the name/term
'series' will be reserved for the series class (and instances) itself
and when a series is in a frame it will be referred to as a column.

Also, added some comments.